Fix #33
So it's a really tricky issue and it's quite complicated to reproduce...
Sometime on Debian 12 we can have the domain which resolv to Link-local address of an interface.
So by example some case we can have have this:

$ ping domain.tld
PING domain.tld(domain.tld (fe80::216:3eff:fef3:11b2%eth0)) 56 data bytes
64 bytes from domain.tld (fe80::216:3eff:fef3:11b2%eth0): icmp_seq=1 ttl=64 time=0.077 ms

But if we use something like host to have the host we have this:

$ host domain.tld
domain.tld has address 127.0.0.1
domain.tld has IPv6 address ::1
domain.tld has IPv6 address 1111:2222:3333::4
domain.tld mail is handled by 5 domain.tld.

And it seem that sogo don't like to end with an ipv6 link local address
So the workaround is to force to use the ipv4 local address and allow the an invalid certificate